refactor: feedback 반영
refactor: value 클래스로 변경

docs: 주석을 제거한다

refactor: object 클래스로 변경한다

refactor: 최대위치 누적계산을 변경한다

refactor: 우승자 객체를 제거한다

refactor: StringBuilder를 제거한다

refactor: String Utils를 제거한다
